[deckbuilder] add "export all" button for user 1
