+++ client/package.json
@@ -10,9 +10,11 @@
   "dependencies": {
     "@types/webpack-env": "^1.15.2",
     "axios": "^0.19.2",
+    "jszip": "^3.5.0",
     "path-to-regexp": "^6.1.0",
     "tslib": "^2.0.0",
     "vue": "^2.6.11",
+    "vue-async-computed": "^3.8.2",
     "vue-router": "^3.2.0",
     "vuex": "^3.4.0"
   }
