+++ client/src/main.ts
@@ -3,8 +3,11 @@ import Vue from 'vue';
 import App from './App.vue';
 import router from './router';
 import { rootStore } from './state/store';
+import AsyncComputed from "vue-async-computed";
 
 Vue.config.productionTip = false;
+
+Vue.use(AsyncComputed);
 
 new Vue({
   router,
